requirements.txt
elasticsearch==5.1.0
python-arango==3.11.0
falcon==1.1.0
falcon-auth==1.1.0
falcon-cors==1.1.2
gunicorn==19.6.0
PyJWT==1.4.2
itsdangerous==0.24
passlib==1.7.1
json-logging-py==0.2
layeredconfig==0.3.2
